Feature Item Details Page
From the menu page- users can select to view item details of said item.

Once in the page they are able to see the details of the page too.

A favourite button is hidden that only shows if a user is logged in.
If they click the button then the dish will be saved to their favourites and
saved on the database.

Requires CSS improvement for future installments.