Feature favourites.php added
Feat include:
Returns favourites list of current User Logged in.
Able to  be redirected to the Item details of favourites item
Able to remove favourite item

Feat sends list of favourite to any provided email account
html body included in email

Tweaked header.php which each css style is called accordingly according to which page the user is on.

Dedicated stylesFunctions.php file to redirect the user to which css to use.